# Vectorize `mismatch` for clang-cl for odd element sizes