--- a/benchmarks/src/mismatch.cpp
+++ b/benchmarks/src/mismatch.cpp
@@ -19,13 +19,24 @@ enum class op {
     lexi,
 };
 
-template <class T, op Op>
+struct color {
+    uint16_t h;
+    uint16_t s;
+    uint16_t l;
+
+    bool operator==(const color&) const = default;
+};
+
+constexpr color c1{30000, 40000, 20000};
+constexpr color c2{30000, 40000, 30000};
+
+template <class T, op Op, T MatchVal = T{'.'}, T MismatchVal = T{'x'}>
 void bm(benchmark::State& state) {
-    vector<T, not_highly_aligned_allocator<T>> a(static_cast<size_t>(state.range(0)), T{'.'});
-    vector<T, not_highly_aligned_allocator<T>> b(static_cast<size_t>(state.range(0)), T{'.'});
+    vector<T, not_highly_aligned_allocator<T>> a(static_cast<size_t>(state.range(0)), MatchVal);
+    vector<T, not_highly_aligned_allocator<T>> b(static_cast<size_t>(state.range(0)), MatchVal);
 
     if (state.range(1) != no_pos) {
-        b.at(static_cast<size_t>(state.range(1))) = 'x';
+        b.at(static_cast<size_t>(state.range(1))) = MismatchVal;
     }
 
     for (auto _ : state) {
@@ -45,6 +56,7 @@ BENCHMARK(bm<uint8_t, op::mismatch>)->Apply(common_args);
 BENCHMARK(bm<uint16_t, op::mismatch>)->Apply(common_args);
 BENCHMARK(bm<uint32_t, op::mismatch>)->Apply(common_args);
 BENCHMARK(bm<uint64_t, op::mismatch>)->Apply(common_args);
+BENCHMARK(bm<color, op::mismatch, c1, c2>)->Apply(common_args);
 
 BENCHMARK(bm<uint8_t, op::lexi>)->Apply(common_args); // still optimized without vector algorithms using memcmp
 BENCHMARK(bm<int8_t, op::lexi>)->Apply(common_args); // optimized with vector algorithms only