--- a/stl/inc/xutility
+++ b/stl/inc/xutility
@@ -454,7 +454,15 @@ size_t _Mismatch_vectorized(const void* const _First1, const void* const _First2
     } else if constexpr (_Element_size == 8) {
         return __std_mismatch_8(_First1, _First2, _Count);
     } else {
-        _STL_INTERNAL_STATIC_ASSERT(false); // unexpected size
+        if constexpr ((_Element_size % 8) == 0) {
+            return __std_mismatch_8(_First1, _First2, _Count * (_Element_size / 8)) / (_Element_size / 8);
+        } else if constexpr ((_Element_size % 4) == 0) {
+            return __std_mismatch_4(_First1, _First2, _Count * (_Element_size / 4)) / (_Element_size / 4);
+        } else if constexpr ((_Element_size % 2) == 0) {
+            return __std_mismatch_2(_First1, _First2, _Count * (_Element_size / 2)) / (_Element_size / 2);
+        } else {
+            return __std_mismatch_1(_First1, _First2, _Count * _Element_size) / _Element_size;
+        }
     }
 }
 _STD_END
@@ -5720,7 +5728,7 @@ namespace ranges {
         _It1 _First1, _It2 _First2, iter_difference_t<_It1> _Count, _Pr _Pred, _Pj1 _Proj1, _Pj2 _Proj2) {
         _STL_INTERNAL_CHECK(_Count >= 0);
 #if _USE_STD_VECTOR_ALGORITHMS
-        if constexpr (_Vector_alg_in_search_is_safe<_It1, _It2, _Pr> && is_same_v<_Pj1, identity>
+        if constexpr (_Equal_memcmp_is_safe<_It1, _It2, _Pr> && is_same_v<_Pj1, identity>
                       && is_same_v<_Pj2, identity>) {
             if (!_STD is_constant_evaluated()) {
                 constexpr size_t _Elem_size = sizeof(iter_value_t<_It1>);